Extracted FsTabBarBase
