+++ src/FlagstoneUI.Core/Controls/Shell/FsTabBar.xaml
@@ -1,7 +1,8 @@
 <?xml version="1.0" encoding="utf-8"?>
 
-<ContentView xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
+<controls:FsTabBarBase xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
              xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
+             xmlns:controls="clr-namespace:FlagstoneUI.Core.Controls"
              VerticalOptions="Start"
              HeightRequest="70"
              x:Name="TabBarRoot"
@@ -23,4 +24,4 @@
                 HorizontalOptions="Fill"
                 x:Name="TabLayout" />
   </Grid>
-</ContentView>
+</controls:FsTabBarBase>
